factor RC into own file
